@@ -1,7 +1,14 @@
 {
   "name": "oryon",
-  "version": "3.0.0",
   "private": true,
+  "version": "0.0.1",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
+  },
   "dependencies": {
     "@auth0/auth0-spa-js": "^1.19.2",
     "@babel/plugin-proposal-private-property-in-object": "^7.14.5",
@@ -27,15 +34,14 @@
     "echarts-for-react": "^3.0.2",
     "formik": "^2.2.9",
     "json-server": "^0.17.4",
-    "jsonwebtoken": "^8.5.1",
     "jwt-decode": "^3.1.2",
     "lodash": "^4.17.19",
     "merge": "^2.1.1",
     "moment": "^2.30.1",
     "notistack": "^2.0.3",
     "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
     "react-material-ui-form-validator": "^3.0.0",
     "react-moment": "^1.1.3",
     "react-perfect-scrollbar": "^1.5.8",
@@ -47,32 +53,16 @@
     "shortid": "^2.2.16",
     "yup": "^0.32.11"
   },
-  "scripts": {
-    "start": "react-scripts --openssl-legacy-provider start",
-    "server": "json-server --watch --id code ./src/fake-db/stores.json --port 3001",
-    "build": "GENERATE_SOURCEMAP=false react-scripts --openssl-legacy-provider build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "ghp": "react-scripts build && gh-pages -d build",
-    "winBuild": "set \"GENERATE_SOURCEMAP=false\" && react-scripts --openssl-legacy-provider build"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ],
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.14.5",
-    "@babel/runtime": "^7.10.5",
+    "@eslint/js": "^9.25.0",
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.2",
+    "@vitejs/plugin-react": "^4.4.1",
     "axios-mock-adapter": "^1.18.2",
-    "babel-plugin-import": "^1.13.0",
-    "cross-env": "^7.0.3",
-    "customize-cra": "^1.0.0",
-    "gh-pages": "^3.2.3",
-    "react-app-rewired": "^2.1.8"
+    "eslint": "^9.25.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.19",
+    "globals": "^16.0.0",
+    "vite": "^6.3.5"
   }
 }
