@@ -1,19 +1,20 @@
 import { CssBaseline } from '@mui/material';
 import { StyledEngineProvider } from '@mui/styled-engine';
 import 'perfect-scrollbar/css/perfect-scrollbar.css';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './app/App';
 import * as serviceWorker from './serviceWorker';
 
-ReactDOM.render(
+
+const root = ReactDOM.createRoot(document.getElementById('root')); // Cambia a createRoot
+root.render(
   <StyledEngineProvider injectFirst>
     <BrowserRouter>
       <CssBaseline />
       <App />
     </BrowserRouter>
-  </StyledEngineProvider>,
-  document.getElementById('root')
+  </StyledEngineProvider>
 );
 
 // for IE-11 support un-comment cssVars() and it's import in this file
