@@ -10,7 +10,7 @@ import {
   ThemeProvider,
   useTheme,
 } from '@mui/material';
-import useSettings from 'app/hooks/useSettings';
+import useSettings from '@/app/hooks/useSettings';
 import { Fragment, useState } from 'react';
 import Scrollbar from 'react-perfect-scrollbar';
 import { themeShadows } from '../Theme/themeColors';
