@@ -1,9 +1,9 @@
 import { Hidden, Switch } from '@mui/material';
 import { Box, styled, useTheme } from '@mui/system';
-import { themeShadows } from 'app/components/Theme/themeColors';
-import useSettings from 'app/hooks/useSettings';
-import { sidenavCompactWidth, sideNavWidth } from 'app/utils/constant';
-import { convertHexToRGB } from 'app/utils/utils';
+import { themeShadows } from '@/app/components/Theme/themeColors';
+import useSettings from '@/app/hooks/useSettings';
+import { sidenavCompactWidth, sideNavWidth } from '@/app/utils/constant';
+import { convertHexToRGB } from '@/app/utils/utils';
 import React from 'react';
 import Brand from '../../Brand';
 import Sidenav from '../../Sidenav';
