@@ -1,10 +1,10 @@
 import { Avatar, Hidden, Icon, IconButton, MenuItem, useMediaQuery } from '@mui/material';
 import { Box, styled, useTheme } from '@mui/system';
-import { UserMenu } from 'app/components';
-import { themeShadows } from 'app/components/Theme/themeColors';
-import useAuth from 'app/hooks/useAuth';
-import useSettings from 'app/hooks/useSettings';
-import { topBarHeight } from 'app/utils/constant';
+import { UserMenu } from '@/app/components';
+import { themeShadows } from '@/app/components/Theme/themeColors';
+import useAuth from '@/app/hooks/useAuth';
+import useSettings from '@/app/hooks/useSettings';
+import { topBarHeight } from '@/app/utils/constant';
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { Span } from '../../Typography';
