@@ -1,4 +1,4 @@
-import useSettings from 'app/hooks/useSettings';
+import useSettings from '@/app/hooks/useSettings';
 import SecondarySidenavTheme from '../Theme/SecondarySidenavTheme/SecondarySidenavTheme';
 import SecondarySidebarContent from './SecondarySidebarContent';
 import SecondarySidebarToggle from './SecondarySidebarToggle';
