@@ -1,5 +1,5 @@
 import { useContext } from 'react'
-import AuthContext from 'app/contexts/JWTAuthContext'
+import AuthContext from '@/app/contexts/JWTAuthContext'
 
 const useAuth = () => useContext(AuthContext)
 
