@@ -1,5 +1,5 @@
 import { useContext } from 'react'
-import SettingsContext from 'app/contexts/SettingsContext'
+import SettingsContext from '@/app/contexts/SettingsContext'
 
 const useSettings = () => useContext(SettingsContext)
 
