@@ -1,9 +1,8 @@
 import AuthGuard from './auth/AuthGuard';
 
-import dashboardRoutes from 'app/views/dashboard/DashboardRoutes';
-
-import NotFound from 'app/views/sessions/NotFound';
-import sessionRoutes from 'app/views/sessions/SessionRoutes';
+import dashboardRoutes from '@/app/views/dashboard/DashboardRoutes';
+import NotFound from '@/app/views/sessions/NotFound';
+import sessionRoutes from '@/app/views/sessions/SessionRoutes';
 import { Navigate } from 'react-router-dom';
 import Layout from './components/Layout/Layout';
 
