@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 import { Box, Card, Grid, Icon, IconButton, styled, Tooltip } from '@mui/material';
-import { Small } from 'app/components/Typography';
+import { Small } from '@/app/components/Typography';
 
 const StyledCard = styled(Card)(({ theme }) => ({
   display: 'flex',
