@@ -1,4 +1,4 @@
-import Loadable from 'app/components/Loadable';
+import Loadable from '@/app/components/Loadable';
 import { lazy } from 'react';
 
 const NotFound = Loadable(lazy(() => import('./NotFound')));
