@@ -1,4 +1,4 @@
-export const API_URL = process.env.REACT_APP_BACKEND_URL || 'http://localhost:3001'
+export const API_URL = import.meta.env.VITE_BACKEND_URL || 'http://localhost:3001'
 
 export const ROLES = {
     Admin: 1,
